extract helper functions into utils
